Tests/LibWeb/Text/input/css/style-invalidation/viewport-resize-shadow-broad-invalidation.html

<!DOCTYPE html>
<script src="../../include.js"></script>
<script>
    function animationFrame() {
        return new Promise(resolve => requestAnimationFrame(resolve));
    }

    async function resizeIframe(iframe, child, width) {
        iframe.style.width = `${width}px`;
        document.body.offsetWidth;
        await animationFrame();
        child.document.body.offsetWidth;
        child.internals.updateStyle();
    }

    asyncTest(async done => {
        const iframe = document.createElement("iframe");
        iframe.style.border = "0";
        iframe.style.width = "300px";
        iframe.style.height = "200px";

        const loaded = new Promise(resolve => iframe.addEventListener("load", resolve, { once: true }));
        iframe.srcdoc = `
            <!DOCTYPE html>
            <x-property-host id="host">host</x-property-host>
        `;
        document.body.appendChild(iframe);
        await loaded;

        const child = iframe.contentWindow;
        const host = child.document.getElementById("host");
        const shadowRoot = host.attachShadow({ mode: "open" });
        const sheet = new child.CSSStyleSheet();
        sheet.replaceSync(`
            :host {
                color: rgb(255, 0, 0);
            }

            @media (min-width: 400px) {
                @property --shadow-media-broad {
                    syntax: '<color>';
                    inherits: false;
                    initial-value: rgb(0, 128, 0);
                }
            }
        `);
        shadowRoot.adoptedStyleSheets = [sheet];
        shadowRoot.appendChild(child.document.createElement("slot"));

        child.internals.updateStyle();
        println(`initial shadow host color: ${child.getComputedStyle(host).color}`);

        child.internals.resetStyleInvalidationCounters();
        await resizeIframe(iframe, child, 500);
        let counters = child.internals.getStyleInvalidationCounters();
        println(`resized shadow host color: ${child.getComputedStyle(host).color}`);
        println(`resized recomputed host side: ${counters.elementStyleRecomputations > 1}`);

        child.internals.resetStyleInvalidationCounters();
        await resizeIframe(iframe, child, 300);
        counters = child.internals.getStyleInvalidationCounters();
        println(`resized back shadow host color: ${child.getComputedStyle(host).color}`);
        println(`resized back recomputed host side: ${counters.elementStyleRecomputations > 1}`);

        done();
    });
</script>
